Logic for study session
We have useEffect - it will run after clicking the btn(cuz isStudyActive) and everytime the the state(studyTime) is changed. In each "iteration" the studyTime is increased by 1, so at some point the if condition will be false, and then we need to reset it to orgianl state. The time left to the endof the session is calculated as times.studyTime(user determines how long the study session has to be) - studyTimetime. I have figured out that it is the easiet way to implement that